requirements.txt
nemreader
psycopg2-binary
python-dotenv
pytz